refactor: add FormSubmission component
