+++ components/Common/form/components/FormSubmission/index.js
@@ -1,0 +1,1 @@
+export { default as FormSubmission } from './FormSubmission'
